+++ mysettingsdialog.py
@@ -8,10 +8,13 @@ store,export,import,iter through scripts
 
 from packages.pymybase.myloggingbase import MyLoggingBase
 import tkinter as tk
+import tkinter.ttk as ttk
+#from idlelib.tooltip import ToolTip
 from tkinter.simpledialog import Dialog
 
 class MySettingsDialog(Dialog,MyLoggingBase):
     
+    result = None
     __inputs = None
     __entries = None
     
@@ -19,7 +22,7 @@ class MySettingsDialog(Dialog,MyLoggingBase):
                  inputs=None,
                  name=None):
         MyLoggingBase.__init__(self,name=name)
-        self.__inputs = inputs
+        self.__inputs = inputs.copy()
         
         Dialog.__init__(self,parent=parent,title=title)
         
@@ -38,19 +41,50 @@ class MySettingsDialog(Dialog,MyLoggingBase):
         except tk.TclError: pass # in ubuntu: _tkinter.TclError: bitmap /.../... not defined
         
         # add inputs
-        self.__entries = list()
+        self.__entries = dict()
+        first_key = None
         r = 0
         for k,v in self.__inputs.items():
-            tk.Label(master, text=k+':',justify=tk.LEFT).grid(row=r,column=0,sticky=tk.W)
-            self.__entries.append(tk.Entry(master))
-            self.__entries[r].grid(row=r,column=1)
+            # get first entry added so we can set that to have focus
+            if first_key is None: first_key = k
             
-            if v is not None:
-                self.__entries[r].insert(0,v)
+            # create label
+            ttk.Label(master, text=k+':',justify=tk.LEFT).grid(row=r,column=0,sticky=tk.W)
             
+            # create entry
+            e = ttk.Entry(master,width=9)
+            e.grid(row=r,column=1)
+            # set default input if any
+            if v is not None: e.insert(0, v)
+            # add to list
+            self.__entries[k] = e
+            
+            # next row!
             r += 1
         
+        # set focus to first entry
+        if self.__entries: return self.__entries[first_key]
+    
+    def buttonbox(self):
+        '''add standard button box.
+
+        override if you do not want the standard buttons
+        '''
+        #Dialog.buttonbox(self)
         
+        # override to use ttk buttons!
+        box = ttk.Frame(self)
+
+        w = ttk.Button(box, text="OK", width=10, command=self.ok, default=tk.ACTIVE)
+        w.pack(side=tk.LEFT, padx=5, pady=5)
+        w = ttk.Button(box, text="Cancel", width=10, command=self.cancel)
+        w.pack(side=tk.LEFT, padx=5, pady=5)
+
+        self.bind("<Return>", self.ok)
+        self.bind("<Escape>", self.cancel)
+
+        box.pack()
+     
     def validate(self):
         '''validate the data
 
@@ -59,18 +93,30 @@ class MySettingsDialog(Dialog,MyLoggingBase):
         '''
         if Dialog.validate(self) == 0: return 0
         
-        # try
-        try:
-            1/1
-        except Exception as e:  #pylint: disable=broad-except
-            tk.messagebox.showwarning(
-                e.__class__.__name__,
-                (e.message if hasattr(e,'message') else e.__class__.__name__) + "\nPlease try again",
-                parent = self
-            )
-            return 0
-        else:
-            return 1
+        for k,entry in self.__entries.items():
+            s = entry.get()
+            
+            # validate if needed
+            if self.__inputs[k] is not None and self.__inputs[k] != s:
+                # validate data type
+                t = type(self.__inputs[k])
+                try:
+                    v = t(s)
+                except ValueError as e:
+                    # failed!
+                    tk.messagebox.showwarning(
+                        e.__class__.__name__,
+                        "'{0}' must be of type '{1}'!\nPlease try again".format(k,t.__name__),
+                        parent = self
+                    )
+                    return 0
+                else:
+                    # good to go, override default with value
+                    self.__inputs[k] = v
+        
+        # passed all validation! - 'apply' to result
+        self.result = self.__inputs
+        return 1
         
     def apply(self):
         '''process the data