+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - myinputdata.py class to (de) serialize each input line
 - Ctrl+Shift+R stops recording by raising event!
 - Run prints actions to console
+- run options! Edit -> Options and prompted on open 
 ## Changed
 - refactored following pylint
 - pymybase located in packages module